Fixed hex grid tiling.  Now the grid is based on three things:  1. a list of x/y pairs called the *unit-hex*  1. a hex-size-multiplier  1. the dimensions of the sdl window
The *unit-hex* is the smallest specification of the polygon to make a hex, and is scaled up based upon hex-size-multiplier to get larger hexes.

Tiling is then done across the sdl window.  This leaves some ragged edges.

In retrospect, the *unit-hex* can probably be replaced by a single (width height) pair.